# google-benchmark 1.9.2 and other cleanups