--- a/benchmarks/inc/utility.hpp
+++ b/benchmarks/inc/utility.hpp
@@ -5,17 +5,12 @@
 
 #include <algorithm>
 #include <cstddef>
-#include <cstdint>
 #include <random>
 #include <vector>
 
-#include <xoshiro.hpp>
-
 template <class Contained>
 std::vector<Contained> random_vector(size_t n) {
-    std::random_device rd;
-    std::uniform_int_distribution<std::uint64_t> id64;
-    xoshiro256ss prng{id64(rd), id64(rd), id64(rd), id64(rd)};
+    std::mt19937_64 prng;
 
     std::vector<Contained> res(n);
 
@@ -25,7 +20,7 @@ std::vector<Contained> random_vector(size_t n) {
 // but is insufficient for aggregate<Data> or non_trivial<Data>.
 #pragma warning(push)
 #pragma warning(disable : 4244) // warning C4244: conversion from 'uint64_t' to 'Data', possible loss of data
-    std::generate(res.begin(), res.end(), [&prng] { return static_cast<Contained>(prng.next()); });
+    std::generate(res.begin(), res.end(), [&prng] { return static_cast<Contained>(prng()); });
 #pragma warning(pop)
 
     return res;