--- a/benchmarks/src/locale_classic.cpp
+++ b/benchmarks/src/locale_classic.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 // GH-3048 <locale>: Double-checked locking for locale::classic
-static void BM_locale_classic(benchmark::State& state) {
+void BM_locale_classic(benchmark::State& state) {
     for (auto _ : state) {
         benchmark::DoNotOptimize(locale::classic());
     }