--- a/benchmarks/src/priority_queue_push_range.cpp
+++ b/benchmarks/src/priority_queue_push_range.cpp
@@ -14,58 +14,56 @@
 #include <vector>
 using namespace std;
 
-namespace {
-    constexpr size_t vec_size = 10'000;
+constexpr size_t vec_size = 10'000;
 
-    template <class T, class Fn>
-    auto create_vec(Fn transformation) {
-        vector<T> vec(vec_size);
-        for (mt19937_64 rnd(1); auto& e : vec) {
-            e = transformation(rnd());
+template <class T, class Fn>
+auto create_vec(Fn transformation) {
+    vector<T> vec(vec_size);
+    for (mt19937_64 rnd(1); auto& e : vec) {
+        e = transformation(rnd());
+    }
+    return vec;
+}
+
+template <class T>
+T cast_to(uint64_t val) {
+    return static_cast<T>(val);
+}
+
+const auto vec_u8     = create_vec<uint8_t>(cast_to<uint8_t>);
+const auto vec_u16    = create_vec<uint16_t>(cast_to<uint16_t>);
+const auto vec_u32    = create_vec<uint32_t>(cast_to<uint32_t>);
+const auto vec_u64    = create_vec<uint64_t>(cast_to<uint64_t>);
+const auto vec_float  = create_vec<float>(cast_to<float>);
+const auto vec_double = create_vec<double>(cast_to<double>);
+
+const auto vec_str  = create_vec<string>([](uint64_t val) { return to_string(static_cast<uint32_t>(val)); });
+const auto vec_wstr = create_vec<wstring>([](uint64_t val) { return to_wstring(static_cast<uint32_t>(val)); });
+
+template <class T, const auto& Data>
+void BM_push_range(benchmark::State& state) {
+    const size_t frag_size = static_cast<size_t>(state.range(0));
+
+    for (auto _ : state) {
+        priority_queue<T> que;
+        span spn{Data};
+
+        while (!spn.empty()) {
+            const size_t take_size = min(spn.size(), frag_size);
+            que.push_range(spn.first(take_size));
+            spn = spn.subspan(take_size);
         }
-        return vec;
+        benchmark::DoNotOptimize(que);
     }
+}
 
-    template <class T>
-    T cast_to(uint64_t val) {
-        return static_cast<T>(val);
-    }
-
-    const auto vec_u8     = create_vec<uint8_t>(cast_to<uint8_t>);
-    const auto vec_u16    = create_vec<uint16_t>(cast_to<uint16_t>);
-    const auto vec_u32    = create_vec<uint32_t>(cast_to<uint32_t>);
-    const auto vec_u64    = create_vec<uint64_t>(cast_to<uint64_t>);
-    const auto vec_float  = create_vec<float>(cast_to<float>);
-    const auto vec_double = create_vec<double>(cast_to<double>);
-
-    const auto vec_str  = create_vec<string>([](uint64_t val) { return to_string(static_cast<uint32_t>(val)); });
-    const auto vec_wstr = create_vec<wstring>([](uint64_t val) { return to_wstring(static_cast<uint32_t>(val)); });
-
-    template <class T, const auto& Data>
-    void BM_push_range(benchmark::State& state) {
-        const size_t frag_size = static_cast<size_t>(state.range(0));
-
-        for (auto _ : state) {
-            priority_queue<T> que;
-            span spn{Data};
-
-            while (!spn.empty()) {
-                const size_t take_size = min(spn.size(), frag_size);
-                que.push_range(spn.first(take_size));
-                spn = spn.subspan(take_size);
-            }
-            benchmark::DoNotOptimize(que);
-        }
-    }
-
-    template <size_t L>
-    void putln(const benchmark::State&) {
-        static bool b = [] {
-            puts("");
-            return true;
-        }();
-    }
-} // namespace
+template <size_t L>
+void putln(const benchmark::State&) {
+    static bool b = [] {
+        puts("");
+        return true;
+    }();
+}
 
 #define TEST_PUSH_RANGE(T, source)      \
     BENCHMARK(BM_push_range<T, source>) \