--- a/benchmarks/src/random_integer_generation.cpp
+++ b/benchmarks/src/random_integer_generation.cpp
@@ -7,7 +7,7 @@
 
 /// Test URBGs alone
 
-static void BM_mt19937(benchmark::State& state) {
+void BM_mt19937(benchmark::State& state) {
     std::mt19937 gen;
     for (auto _ : state) {
         benchmark::DoNotOptimize(gen());
@@ -15,7 +15,7 @@ static void BM_mt19937(benchmark::State& state) {
 }
 BENCHMARK(BM_mt19937);
 
-static void BM_mt19937_64(benchmark::State& state) {
+void BM_mt19937_64(benchmark::State& state) {
     std::mt19937_64 gen;
     for (auto _ : state) {
         benchmark::DoNotOptimize(gen());
@@ -23,7 +23,7 @@ static void BM_mt19937_64(benchmark::State& state) {
 }
 BENCHMARK(BM_mt19937_64);
 
-static void BM_lcg(benchmark::State& state) {
+void BM_lcg(benchmark::State& state) {
     std::minstd_rand gen;
     for (auto _ : state) {
         benchmark::DoNotOptimize(gen());
@@ -37,11 +37,11 @@ std::uint32_t GetMax() {
     return dist(gen);
 }
 
-static const std::uint32_t maximum = GetMax(); // random divisor to prevent strength reduction
+const std::uint32_t maximum = GetMax(); // random divisor to prevent strength reduction
 
 /// Test mt19937
 
-static void BM_raw_mt19937_old(benchmark::State& state) {
+void BM_raw_mt19937_old(benchmark::State& state) {
     std::mt19937 gen;
     std::_Rng_from_urng<std::uint32_t, decltype(gen)> rng(gen);
     for (auto _ : state) {
@@ -50,7 +50,7 @@ static void BM_raw_mt19937_old(benchmark::State& state) {
 }
 BENCHMARK(BM_raw_mt19937_old);
 
-static void BM_raw_mt19937_new(benchmark::State& state) {
+void BM_raw_mt19937_new(benchmark::State& state) {
     std::mt19937 gen;
     std::_Rng_from_urng_v2<std::uint32_t, decltype(gen)> rng(gen);
     for (auto _ : state) {
@@ -61,7 +61,7 @@ BENCHMARK(BM_raw_mt19937_new);
 
 /// Test mt19937_64
 
-static void BM_raw_mt19937_64_old(benchmark::State& state) {
+void BM_raw_mt19937_64_old(benchmark::State& state) {
     std::mt19937_64 gen;
     std::_Rng_from_urng<std::uint64_t, decltype(gen)> rng(gen);
     for (auto _ : state) {
@@ -70,7 +70,7 @@ static void BM_raw_mt19937_64_old(benchmark::State& state) {
 }
 BENCHMARK(BM_raw_mt19937_64_old);
 
-static void BM_raw_mt19937_64_new(benchmark::State& state) {
+void BM_raw_mt19937_64_new(benchmark::State& state) {
     std::mt19937_64 gen;
     std::_Rng_from_urng_v2<std::uint64_t, decltype(gen)> rng(gen);
     for (auto _ : state) {
@@ -81,7 +81,7 @@ BENCHMARK(BM_raw_mt19937_64_new);
 
 /// Test minstd_rand
 
-static void BM_raw_lcg_old(benchmark::State& state) {
+void BM_raw_lcg_old(benchmark::State& state) {
     std::minstd_rand gen;
     std::_Rng_from_urng<std::uint32_t, decltype(gen)> rng(gen);
     for (auto _ : state) {
@@ -90,7 +90,7 @@ static void BM_raw_lcg_old(benchmark::State& state) {
 }
 BENCHMARK(BM_raw_lcg_old);
 
-static void BM_raw_lcg_new(benchmark::State& state) {
+void BM_raw_lcg_new(benchmark::State& state) {
     std::minstd_rand gen;
     std::_Rng_from_urng_v2<std::uint32_t, decltype(gen)> rng(gen);
     for (auto _ : state) {