--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -417,8 +417,7 @@ auto _Max_vectorized(_Ty* const _First, _Ty* const _Last) noexcept {
 }
 
 template <size_t _Element_size>
-inline size_t // TRANSITION, GH-4496
-    _Mismatch_vectorized(const void* const _First1, const void* const _First2, const size_t _Count) noexcept {
+size_t _Mismatch_vectorized(const void* const _First1, const void* const _First2, const size_t _Count) noexcept {
     if constexpr (_Element_size == 1) {
         return __std_mismatch_1(_First1, _First2, _Count);
     } else if constexpr (_Element_size == 2) {