… NOTICE.txt
… benchmarks/inc/xoshiro.hpp
… stl/CMakeLists.txt